Refactor: Did a couple of bigger changes
New idenfitier for the Tree. This is now an enum describing the level
and the contents of the tree better. This will allow for a calling and methods
and getting properties.

The tree binary has been moved to examples. It was not a useful program, and it shadows
the ubiquitous tree that show a file tree.

Added logging with the tracing crate. This is helpful when debugging issues in the tui

Updated Cargo.lock