Method calls! You can now do dbus method calls in the TUI!
When hitting Enter on a method in the tree a popup will show.
You can then enter each argument in text fields in this popup
the arguments are then parsed into dbus types that are sent with the method call
The return value is then put into the output text box.

To do this a parser for a human readable format for dbus types was created
parsers are created for a given type signature, and the then input is validated against that
The parser is written using chumsky with subparsers for each dbus type.