Giant refactoring with multiple new features
Refactors the application into a components structure that comes from
the ratatui documentation.
A component is a part of the application that is rendered and receives
actions as whenever something happens. These Actions are created from
KeyEvents, or responses from dBus.

The components are:
  - Service view
  - Objects view
  - Result view
  - Call view

The Service view shows the current services, and you can select which
service to show the objects
The Objects view, will show the objects of the service you selected, you
can then set or get properties or call methods.
The result view will show the results from calls or properties, as well
as errors
The Call view is for allowing method calls or setting properties, the
arguments are shown and you can enter a value that will be parsed

The new features are:
 - Results view
 - Help view
 - Getting and setting properties
 - hjkl navigation